+++ CarlTour/CTTourSelectorViewController.m
@@ -28,7 +28,7 @@
     return self;
 }
 
-- (int) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return [self.tours count];
 }
